A metric shit-ton of edits.
It's not that I'm too lazy to write an edit summary, but I changed a lot of shit to get stuff working that I fucked up before.

Try and go through the changes to make sense of what I did, because I did some funky stuff in the new Vader file and such to get things working. I know I also did a few spacing changes....