Update VaderAssistsLand.lua
Revising. Localizing some variables, and moving around some. It may not be totally consistent but none of this needs to be outside of the ``while true do`` scope, so it's cleanest to keep it there.

Also moving the check to see if Vader's available, which will save 0.0000587293437492nd of a frame or 0.0148721kb of memory by exiting out faster if Vader isn't available.